+++ src/gui/widgets/timeline/compound_timeline_widget.hpp
@@ -52,6 +52,7 @@ private slots:
     void custom_context_menu(const QPoint& p);
     void add_keyframe();
     void remove_keyframe();
+    void remove_all_keyframes();
     void on_scroll(int amount);
     void keyframe_action_enter();
     void keyframe_action_exit();